Move file management from TextView to FileExplorer & FileTabs
