+++ src/components/tab/FileTabs.vue
@@ -42,10 +42,12 @@ import {
 } from 'vue';
 import FileTabHeader from 'src/components/tab/FileTabHeader.vue';
 import FileEvent from 'src/composables/events/FileEvent';
+import GitEvent from 'src/composables/events/GitEvent';
+import { exists, getStatus } from 'src/composables/Project';
 
 const props = defineProps({
-  fileInformations: {
-    type: Object,
+  projectName: {
+    type: String,
     required: true,
   },
 });
@@ -54,6 +56,14 @@ const fileTabArray = ref([]);
 const activeFileId = ref(null);
 
 let selectFileNodeSubscription;
+let deleteFileSubscription;
+let updateEditorContentSubscription;
+let addRemoteSubscription;
+let checkoutSubscription;
+let pullSubscription;
+let addFileSubscription;
+let commitFilesSubscription;
+let globalUploadFilesEventSubscription;
 
 /**
  * Update activeFileId by setting its id equal to the last element of fileTabArray,
@@ -75,9 +85,9 @@ function setLastFileActive() {
  * @param {String} fileId - Id of closed file.
  */
 function deleteFileTab(fileId) {
-  fileTabArray.value = fileTabArray.value.filter(({ id }) => id !== fileId);
+  fileTabArray.value = fileTabArray.value.filter(({ id }) => !id.startsWith(fileId));
 
-  if (fileId === activeFileId.value) {
+  if (activeFileId.value && activeFileId.value.startsWith(fileId)) {
     setLastFileActive();
   }
 }
@@ -88,40 +98,80 @@ function deleteFileTab(fileId) {
  */
 function onSelectFileNode(node) {
   if (!fileTabArray.value.some(({ id }) => id === node.id)) {
-    fileTabArray.value.push({
-      id: node.id,
-      label: node.label,
-      information: props.fileInformations.find(({ path }) => path === node.id),
-    });
+    fileTabArray.value.push(node);
   }
 
   activeFileId.value = node.id;
 }
 
 /**
- * Update fileTabArray when props.fileInformations is updated.
- * If the previous selected file no longer exists, update activeFileId.
+ * Update status of the fileTab corresponding to the given parameter.
+ * @param {String} filePath - Path correponding to the fileTab id.
+ * @returns {Promise<void>} Promise with nothing on success otherwise an error.
  */
-function updateFileTabs() {
-  const projectFilesIds = props.fileInformations.map((file) => file.path);
-  fileTabArray.value = fileTabArray.value
-    .filter(({ id }) => projectFilesIds.includes(id))
-    .map((fileTab) => ({
-      ...fileTab,
-      information: props.fileInformations.find(({ path }) => path === fileTab.id),
-    }));
+async function updateFileStatus(filePath) {
+  if (fileTabArray.value.length === 0) {
+    return;
+  }
 
-  const isActiveFileInFiles = fileTabArray.value
-    .find(({ id }) => id === activeFileId.value);
+  const [fileStatus] = await getStatus(props.projectName, [filePath], (path) => path === filePath);
+  const index = fileTabArray.value.findIndex((fileTab) => fileTab.id === filePath);
 
-  if (!isActiveFileInFiles) {
-    setLastFileActive();
+  if (index !== -1) {
+    fileTabArray.value[index].information = fileStatus;
   }
 }
 
-watch(() => props.fileInformations, () => {
-  updateFileTabs();
-}, { deep: true });
+/**
+ * Update status of all the fileTabs corresponding to the given parameter.
+ * @param {String[]} allFilePaths - Array of path correponding to the id of the fileTabs.
+ * @returns {Promise<void>} Promise with nothing on success otherwise an error.
+ */
+async function updateAllFilesStatus(allFilePaths) {
+  if (fileTabArray.value.length === 0) {
+    return;
+  }
+
+  // update return of GitEvent.CommitEvent
+  // const allFilePaths = event.map(({ path }) => path);
+  const allFileStatus = await getStatus(
+    props.projectName,
+    allFilePaths,
+    (path) => allFilePaths.includes(path),
+  );
+
+  allFileStatus.forEach((fileStatus) => {
+    const index = fileTabArray.value.findIndex((fileTab) => fileTab.id === fileStatus.path);
+
+    if (index !== -1) {
+      fileTabArray.value[index].information = fileStatus;
+    }
+  });
+}
+
+/**
+ * Update fileTab array and all there status.
+ * @returns {Promise<void>} Promise with nothing on success otherwise an error.
+ */
+async function updateAllFileTabs() {
+  await Promise.allSettled(
+    fileTabArray.value.map(
+      (fileTab) => exists(`${props.projectName}/${fileTab.id}`).then((exist) => ({ fileTab, exist })),
+    ),
+  ).then((results) => {
+    fileTabArray.value = results
+      .filter(({ status, value }) => status === 'fulfilled' && value.exist)
+      .map(({ value }) => value.fileTab);
+  });
+
+  setLastFileActive();
+
+  await updateAllFilesStatus(
+    fileTabArray.value.map(
+      ({ id }) => id,
+    ),
+  );
+}
 
 watch(activeFileId, () => {
   FileEvent.SelectFileTabEvent.next(activeFileId.value);
@@ -129,10 +179,29 @@ watch(activeFileId, () => {
 
 onMounted(() => {
   selectFileNodeSubscription = FileEvent.SelectFileNodeEvent.subscribe(onSelectFileNode);
+  deleteFileSubscription = FileEvent.DeleteFileEvent.subscribe(({ id }) => deleteFileTab(id));
+  updateEditorContentSubscription = FileEvent.UpdateEditorContentEvent.subscribe(updateFileStatus);
+  addRemoteSubscription = GitEvent.AddRemoteEvent.subscribe(updateAllFileTabs);
+  checkoutSubscription = GitEvent.CheckoutEvent.subscribe(updateAllFileTabs);
+  pullSubscription = GitEvent.PullEvent.subscribe(updateAllFileTabs);
+  addFileSubscription = GitEvent.AddEvent.subscribe(updateFileStatus);
+  commitFilesSubscription = GitEvent.CommitEvent.subscribe(
+    (event) => updateAllFilesStatus(event.map(({ path }) => path)),
+  );
+  globalUploadFilesEventSubscription = FileEvent.GlobalUploadFilesEvent
+    .subscribe(updateAllFileTabs);
 });
 
 onUnmounted(() => {
   selectFileNodeSubscription.unsubscribe();
+  deleteFileSubscription.unsubscribe();
+  updateEditorContentSubscription.unsubscribe();
+  addRemoteSubscription.unsubscribe();
+  checkoutSubscription.unsubscribe();
+  pullSubscription.unsubscribe();
+  addFileSubscription.unsubscribe();
+  commitFilesSubscription.unsubscribe();
+  globalUploadFilesEventSubscription.unsubscribe();
 });
 </script>
